Вынес логику вычисления стоимости из маппера в сервис
