+++ src/main/java/ru/bakulin/daily_booking_service/service/impl/BookingServiceImpl.java
@@ -1,6 +1,8 @@
 package ru.bakulin.daily_booking_service.service.impl;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
+import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
@@ -12,6 +14,7 @@ import ru.bakulin.daily_booking_service.dto.BookingDtoRq;
 import ru.bakulin.daily_booking_service.dto.BookingDtoRs;
 import ru.bakulin.daily_booking_service.dto.ClientDto;
 import ru.bakulin.daily_booking_service.dto.PageDto;
+import ru.bakulin.daily_booking_service.entity.Advert;
 import ru.bakulin.daily_booking_service.entity.Booking;
 import ru.bakulin.daily_booking_service.exception.EntityNotFound;
 import ru.bakulin.daily_booking_service.exception.UnavailableBookingPeriod;
@@ -39,19 +42,38 @@ public class BookingServiceImpl implements BookingService {
     ClientDto client = dtoRq.getClient();
 
     if (Objects.nonNull(client.getId()) && !clientRepository.existsById(client.getId())) {
-      throw new EntityNotFound("Клиент с указанным Id= %s не найден в БД".formatted(client.getId()));
+      throw new EntityNotFound(
+          "Клиент с указанным Id= %s не найден в БД".formatted(client.getId()));
     }
 
     if (Objects.isNull(client.getId())) {
       dtoRq.setClient(clientService.save(client));
     }
 
+    BigDecimal resultPrice = getResultPrice(dtoRq);
+
     checkFreeBookingPeriod(dtoRq);
 
     Booking entity = mapper.toEntityWithRelation(dtoRq);
+    entity.setAmount(resultPrice);
+
     Booking booking = repository.save(entity);
 
     return mapper.toDtoRs(booking);
+  }
+
+  protected BigDecimal getResultPrice(BookingDtoRq dto) {
+    Advert advert = advertRepository.findById(dto.getAdvertId()).orElseThrow(
+        () -> new EntityNotFound(
+            "Не найдено Объявление с указанным Id= %s".formatted(dto.getAdvertId())));
+
+    LocalDate start = dto.getDateStart();
+    LocalDate finish = dto.getDateFinish();
+
+    checkCorrectInputPeriodBooking(start, finish);
+    long countDayBooking = ChronoUnit.DAYS.between(start, finish);
+
+    return advert.getPrice().multiply(BigDecimal.valueOf(countDayBooking));
   }
 
   @Override
@@ -63,6 +85,13 @@ public class BookingServiceImpl implements BookingService {
     Page<Booking> page = repository.findAllByClientEmail(email, pageRequest);
 
     return mapper.toPageDto(page);
+  }
+
+  private void checkCorrectInputPeriodBooking(LocalDate start, LocalDate finish) {
+    if (start.isAfter(finish)) {
+      throw new UnavailableBookingPeriod(
+          "Ошибка в указании периода бронирования. Дата окончания бронирования ранее даты начала бронирования.");
+    }
   }
 
   private void checkFreeBookingPeriod(BookingDtoRq dtoRq) {
